backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
httpx>=0.27.0
pydantic-settings>=2.6.0
yt-dlp>=2024.1.0
openai>=1.0.0
google-generativeai>=0.8.0
